Core/DataStores: Updated db2 structures to 9.0.2
Port From (https://github.com/TrinityCore/TrinityCore/commit/b82a3a557a7bf13d48342ea189325550059b622d)

+++ Source/Framework/Constants/GameObjectConst.cs
@@ -17,7 +17,7 @@
 
 namespace Framework.Constants
 {
-    public enum GameObjectTypes : byte
+    public enum GameObjectTypes : int
     {
         Door = 0,
         Button = 1,
